Adjust project's structure.
