+++ .gitignore
@@ -1,7 +1,7 @@
-/.idea
-/*.wav
-/*.bmp
-/*.tiff
-/*.png
-/*.pyc
-/*.tif
+.idea
+*.wav
+*.bmp
+*.tiff
+*.png
+*.pyc
+*.tif
