+++ core/__init__.py
@@ -1,0 +1,3 @@
+# T_T coding=utf-8 T_T
+
+from api import *
